+++ inst/staticexports/str_detect.R
@@ -1,12 +1,15 @@
-str_detect <- function(string, pattern) {
+str_detect <- function(string, pattern, negate = FALSE) {
 	is_fixed <- inherits(pattern, "fixed")
 	ignore.case <- isTRUE(attr(pattern, "options")$case_insensitive)
 
-	grepl(
-		pattern = pattern,
+	indices <- grep(
+		pattern,
 		x = string,
 		ignore.case = ignore.case,
 		perl = !is_fixed,
-		fixed = is_fixed
+		fixed = is_fixed,
+		invert = negate
 	)
+
+	seq_along(string) %in% indices
 }